requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
gunicorn==23.0.0
httpx==0.27.2
python-dotenv==1.0.1
